+++ nodejs/recompile_node.sh
@@ -7,13 +7,22 @@ localpwd=`pwd`
 VER="$@"
 
 echo "Deleting old node version"
+
 rm -R node-*
 
+echo
+echo "Downloading new version from https://nodejs.org..."
+echo
+
+wget https://nodejs.org/dist/v$VER/node-v$VER.tar.gz
+
+echo
 echo
 echo "Untar new version"
 echo
 
-tar zxvf ../archive/node-v$VER.tar.gz
+tar zxvf node-v$VER.tar.gz
+rm node-v$VER.tar.gz
 cd node-v$VER/
 ./configure
 ./configure --prefix=$localpwd/node-$VER/